requirements.txt
requests~=2.32.3
pandas~=2.2.2
python-dotenv~=0.20.0
numpy~=1.26.4
scikit-learn~=1.5.1
matplotlib~=3.7.3
plotly~=5.22.0
plotly-resampler~=0.10.0
seaborn~=0.11.2
fredapi~=0.5.0
yfinance~=0.2.41
fredapi~=0.5.2
statsmodels~=0.14.1
cvxopt~=1.3.2